#29 게시글 등록, 수정을 위한 Dto 클래스의 필드와 어노테이션을 수정
